implement load word

+++ src/headers/n64/mem.h
@@ -13,9 +13,9 @@ struct Mem
 
 void reset_mem(Mem &mem, const std::string &filename);
 
-u8 read_u8(N64 &n64,u64 addr);
-u16 read_u16(N64 &n64,u64 addr);
-u32 read_u32(N64 &n64,u64 addr);
-u64 read_u64(N64 &n64,u64 addr);
+u8 read_u8(N64 &n64,u32 addr);
+u16 read_u16(N64 &n64,u32 addr);
+u32 read_u32(N64 &n64,u32 addr);
+u64 read_u64(N64 &n64,u32 addr);
 
 }
